requirements.txt
alembic==1.9.1
barnum==0.5.1
bcrypt==4.0.1
blinker==1.5
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==4.0.1
Flask-Principal==0.4.0
Flask-Security-Too==5.1.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
future==0.18.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
migrate==0.3.8
mysql-connector-python==8.0.31
passlib==1.7.4
protobuf==3.20.1
SQLAlchemy==1.4.46
Werkzeug==2.2.2
WTForms==3.0.1
